+++ disassemblers/ofrak_binary_ninja/src/ofrak_binary_ninja/components/blocks/unpackers.py
@@ -17,8 +17,7 @@ from ofrak.core.program import Program
 from ofrak.model.component_model import ComponentConfig
 from ofrak.resource import Resource
 from ofrak.service.resource_service_i import ResourceFilter
-from ofrak_binary_ninja.model import BinaryNinjaAnalysisResource
-from ofrak_binary_ninja.model import BinaryNinjaAnalysis
+from ofrak_binary_ninja.model import BinaryNinjaAnalysisResource, BinaryNinjaAnalysis
 
 LOGGER = logging.getLogger(__name__)